salman : Membuat dan membuat tabel di  migrasi foreign_key

+++ database/migrations/2025_03_09_045257_create_siswas_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('parents')->nullable();
             $table->string('contact_parents')->nullable();
             $table->string('foto')->nullable();
-            $table->integer('kelas_id')->nullable();
+            $table->unsignedBigInteger('kelas_id')->nullable();
             $table->timestamps();
         });
     }
