+++ database/migrations/2025_03_10_033752_create_raports_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('raports', function (Blueprint $table) {
             $table->id();
-            $table->integer('siswa_id')->nullable();
+            $table->unsignedBigInteger('siswa_id')->nullable();
             $table->integer('semester')->nullable();
             $table->string('mata_pelajaran')->nullable();
             $table->decimal('nilai')->nullable();
